+++ src/persistence/topic/repository.ts
@@ -3,6 +3,7 @@ import TopicModel from './model';
 import ITopic from '../../domain/topic/type';
 import validateTopic from '../../domain/topic/validator';
 import validateObjectId from '../validator';
+import { topicsCache } from '../cache';
 
 const modelToTopic = (model: { _doc: ITopic; }) => model._doc as ITopic;
 
@@ -14,10 +15,14 @@ const modelsToTopics = (models: { _doc: ITopic; }[]): ITopic[] => {
 };
 
 const searchTopicByName = async (name: string): Promise<ITopic[]> => {
-  let topics: ITopic[] = [];
+  const topics: ITopic[] = [];
   try {
-    const topicModels = await TopicModel.find({ name });
-    topics = modelsToTopics(topicModels);
+    Object.keys(topicsCache).forEach((key) => {
+      const article = topicsCache[key];
+      if (article.name === name) {
+        topics.push(article);
+      }
+    });
   } catch (error: any) {
     logger.error(error);
   }
@@ -26,37 +31,66 @@ const searchTopicByName = async (name: string): Promise<ITopic[]> => {
 
 const getAllTopics = async (): Promise<ITopic[]> => {
   const allTopicModels = await TopicModel.find({});
-  return modelsToTopics(allTopicModels);
+
+  const topicInterfaces = modelsToTopics(allTopicModels);
+  topicInterfaces.forEach((topic) => {
+    topicsCache[topic._id] = topic;
+  });
+
+  return topicInterfaces;
 };
 
 const getTopicById = async (id: string): Promise<ITopic | null> => {
   validateObjectId(id);
 
+  if (topicsCache[id] !== null) return topicsCache[id];
+
   const topicModel = await TopicModel.findById(id);
 
   if (topicModel === null) return null;
-  return modelToTopic(topicModel);
+
+  const topicInterface = modelToTopic(topicModel);
+  topicsCache[topicInterface._id] = topicInterface;
+  return topicInterface;
 };
 
 const createTopic = async (topic: ITopic): Promise<ITopic> => {
   validateTopic(topic);
 
   const _topicModel = await TopicModel.create(topic);
-  return modelToTopic(_topicModel);
+  const topicInterface = modelToTopic(_topicModel);
+  topicsCache[topicInterface._id] = topicInterface;
+  return topicInterface;
 };
 
-const updateTopic = async (topic: ITopic): Promise<ITopic> => {
+const updateTopic = async (topic: ITopic): Promise<ITopic | null> => {
+  let _topic: ITopic;
+  if (topicsCache[topic._id] !== null) {
+    _topic = { ...topicsCache[topic._id] };
+    _topic.name = topic.name;
+    _topic.description = topic.description;
+    _topic.color = topic.color;
+  } else {
+    return null;
+  }
+
   validateTopic(topic);
 
   const _topicModel = await TopicModel.findByIdAndUpdate(topic._id, topic, { new: true });
-  return modelToTopic(_topicModel);
+  const topicInterface = modelToTopic(_topicModel);
+  topicsCache[topicInterface._id] = topicInterface;
+  return topicInterface;
 };
 
-const deleteTopic = async (id: string): Promise<ITopic> => {
+const deleteTopic = async (id: string): Promise<ITopic | null> => {
   validateObjectId(id);
 
+  if (topicsCache[id] === null) return null;
+
   const topicModel = await TopicModel.findByIdAndDelete(id);
-  return modelToTopic(topicModel);
+  const topicInterface = modelToTopic(topicModel);
+  delete topicsCache[topicInterface._id];
+  return topicInterface;
 };
 
 export {
